Change max/min values for jqueryui > 1.11.2
Since jqueryui released version 1.11.2 the pips plugin has been
incompatible with it. This was due to the jqui team (awesomely) changing
the way of calculating stepped-values. They removed the remainders of
sliders with odd step values.

It was possible to handle this in a backwards compatible way by utilising
the ._valueMax() and ._valueMin() methods from the jqui slider plugin, instead
of using the .options.max/.options.min values.