Fix IE issues, Fix Animations, Fix Mouse Events
Firstly I was using .map() which was causing errors in IE7/8

Secondly I noticed that Chrome/Firefox both had buggy CSS anitations on
the pip floats due to a bug in their CSS parsing. So I updated the css to
account for that.

Thirdly there was a few small bugs with the mousedown handling, meaning
that the pip labels were not getting the right events bound to them, and
the UI looked odd when pip-labels were clicked. Fixed this by preventing
the original mouse event from the JQUI library and then deciding to run it
if needed, otherwise using the labelClick() method to handle it. -- No
more glitchy animations, or un-intended pip clicking on sliders with tons
of pips.